mr_ict_backend/requirements.txt
asgiref
Django
djangorestframework
Pillow
pytz
sqlparse
psycopg2-binary
redis
celery
channels-redis
requests
pyfcm
django-cors-headers
daphne
djangorestframework-simplejwt
django-environ
gunicorn
ffmpeg-python
#sentence-transformers==2.2.2 
#huggingface_hub==0.14.1
#faiss-cpu==1.12.0
#cacheout==0.14.1